requirements.txt
# Discord Bot
discord.py>=2.3.2
python-dotenv>=1.0.0

# FastAPI Backend
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Cryptocurrency Payment Processing
bit>=0.8.0  # Bitcoin library
litecoin>=0.1.1  # Litecoin support
requests>=2.31.0

# Database
aiosqlite>=0.19.0
sqlalchemy>=2.0.23
alembic>=1.13.0

# Security & Encryption
cryptography>=41.0.7
pycryptodome>=3.19.0
python-jose[cryptography]>=3.3.0

# OSRS Integration
pyautogui>=0.9.54  # For automation
pillow>=10.1.0  # Image processing
opencv-python>=4.8.1.78  # Computer vision for screen detection

# Webhooks & Tunneling
pyngrok>=7.0.1

# Utilities
aiohttp>=3.9.1
asyncio-throttle>=1.0.2
python-dateutil>=2.8.2
pytz>=2023.3

# Logging & Monitoring
colorlog>=6.8.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
httpx>=0.25.2  # For testing FastAPI

# Development
black>=23.12.0
flake8>=6.1.0
mypy>=1.7.1
